python-service/requirements.txt
fastapi
uvicorn
python-docx
python-pptx
mediapipe
opencv-python-headless
langchain
langchain-google-genai
pydantic
pydantic-settings
python-dotenv
